+++ R/ggbrain-package.R
@@ -1,0 +1,5 @@
+## usethis namespace: start
+#' @importFrom Rcpp sourceCpp
+#' @useDynLib ggbrain, .registration = TRUE
+## usethis namespace: end
+NULL